rag-complaint_bot/requirements.txt
fastapi
uvicorn
pydantic
streamlit
langchain
python-dotenv
sentence-transformers
faiss-cpu
requests
PyPDF2
langchain-community